Cleaning up some boilerplate from app
