+++ public/index.html
@@ -28,24 +28,8 @@
   </head>
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
-
     <script src="https://unpkg.com/survey-react"></script>
-
     <link href="https://unpkg.com/survey-react/survey.min.css" type="text/css" rel="stylesheet"/>
-
-    <div id="surveyElement"></div>
-    <div id="eligibilityResultsContainer"></div>
-
     <div id="root"></div>
-    <!--
-      This HTML file is a template.
-      If you open it directly in the browser, you will see an empty page.
-
-      You can add webfonts, meta tags, or analytics to this file.
-      The build step will place the bundled scripts into the <body> tag.
-
-      To begin the development, run `npm start` or `yarn start`.
-      To create a production bundle, use `npm run build` or `yarn build`.
-    -->
   </body>
 </html>
